move quart_api to quart folder and fix some errors
