+++ pyproject.toml
@@ -81,3 +81,4 @@ prompto_run_checks = "prompto.scripts.run_checks:main"
 prompto_create_judge_file = "prompto.scripts.create_judge_file:main"
 prompto_obtain_missing_jsonl = "prompto.scripts.obtain_missing_jsonl:main"
 prompto_convert_images = "prompto.scripts.convert_images:main"
+prompto_quart_server = "prompto.models.quart.quart_api:main"
